requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
Flask-Babel==3.1.0